[REFACTOR] : Loader 재사용 할 수 있도록 DesginResource로 이동
