+++ core/designsystem/src/main/res/raw/raw_loading.json
