+++ feature/notice/build.gradle.kts
@@ -29,7 +29,6 @@ dependencies {
 
     implementation(libs.bundles.androidx)
     implementation(libs.material)
-    implementation(libs.lottie.compose)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.test.junit)
     androidTestImplementation(libs.androidx.test.espresso)
